requirements.txt
langchain
openai
neo4j
python-dotenv
spacy
faiss-cpu